refactor: organize AI SDK message conversion and tests
