+++ .changeset/tidy-ai-sdk-message-conversion.md
@@ -1,0 +1,5 @@
+---
+'@openai/agents-extensions': patch
+---
+
+refactor: organize AI SDK message conversion and its tests without changing adapter behavior.
